requirements.txt
fastapi
uvicorn[standard]
anyio
google-api-python-client
playwright
pandas
openpyxl
python-dotenv
playwright-stealth
requests
beautifulsoup4
